pipeline/requirements.txt
requests
beautifulsoup4
pymupdf
Pillow
numpy
google.genai
python-dotenv
sqlalchemy
psycopg2-binary
boto3